move user search to users' repo
